+++ tools/ckan-utils/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG ckan-utils
 
+17.January.2024
+- make import more robust, rely on `package_id` of the datasets
+- use Python standard libs for fetching data and unzipping resources
+- dont rely on resource `mimetype`, check the `url` instead
+
 14.January.2024
 - adds `User-Agent` header to `opentransportdata.swiss` API requests
 - adds `fetch_metadata_cli.py` script that fetches JSON metadata from the CKAN API